requirements.txt
numpy
pandas
pymongo
python-dotenv
# -e .